requirements.txt
flask[async]
requests
python-dotenv
aiohttp
gunicorn
beautifulsoup4